Move namespace-override tests to test_standalone2.py
