+++ python/ray/serve/tests/test_standalone2.py
@@ -1,9 +1,20 @@
 import sys
 
 import pytest
+import requests
 
 import ray
 from ray import serve
+from ray.serve.api import internal_get_global_client
+
+
+@pytest.fixture
+def shutdown_ray():
+    if ray.is_initialized():
+        ray.shutdown()
+    yield
+    if ray.is_initialized():
+        ray.shutdown()
 
 
 def test_standalone_actor_outside_serve():
@@ -24,6 +35,44 @@ def test_standalone_actor_outside_serve():
 
     ray.get(a.ready.remote())
     ray.shutdown()
+
+
+@pytest.mark.parametrize("detached", [True, False])
+def test_override_namespace(shutdown_ray, detached):
+    """Test the _override_controller_namespace flag in serve.start()."""
+
+    ray_namespace = "ray_namespace"
+    controller_namespace = "controller_namespace"
+
+    ray.init(namespace=ray_namespace)
+    serve.start(detached=detached, _override_controller_namespace=controller_namespace)
+
+    controller_name = internal_get_global_client()._controller_name
+    ray.get_actor(controller_name, namespace=controller_namespace)
+
+    serve.shutdown()
+
+
+@pytest.mark.parametrize("detached", [True, False])
+def test_deploy_with_overriden_namespace(shutdown_ray, detached):
+    """Test deployments with overriden namespace."""
+
+    ray_namespace = "ray_namespace"
+    controller_namespace = "controller_namespace"
+
+    ray.init(namespace=ray_namespace)
+    serve.start(detached=detached, _override_controller_namespace=controller_namespace)
+
+    for iteration in range(2):
+
+        @serve.deployment
+        def f(*args):
+            return f"{iteration}"
+
+        f.deploy()
+        assert requests.get("http://localhost:8000/f").text == f"{iteration}"
+
+    serve.shutdown()
 
 
 if __name__ == "__main__":
